refactor: replace db types
